Update third-party-software-updates.md
The section Additional requirements when the SUP is remote from the top-level site server confused me until I realized that this SSL cert was to be a self-signed cert. Therefore, I felt that this needed to be added to Point 1 so readers can see this from the offset and therefore point 2 onward makes more sense.